effective-java/src/main/java/study/effectivejava/chap6/item34/Operation.java
package study.effectivejava.chap6.item34;

import lombok.RequiredArgsConstructor;
import lombok.ToString;

import java.util.Map;
import java.util.Objects;
import java.util.Optional;
import java.util.stream.Stream;

import static java.util.stream.Collectors.toMap;

/**
 * 상수별 메서드 구현
 */
@RequiredArgsConstructor
public enum Operation {

    PLUS("+") {
        @Override
        public double apply(double x, double y) {
            return x + y;
        }
    },
    MINUS("-") {
        @Override
        public double apply(double x, double y) {
            return x - y;
        }
    },
    TIMES("*") {
        @Override
        public double apply(double x, double y) {
            return x * y;
        }
    },
    DIVIDE("/") {
        @Override
        public double apply(double x, double y) {
            return x / y;
        }
    };

    private static final Map<String, Operation> stringToEnum = Stream.of(values())
                                                                     .collect(toMap(Objects::toString, e -> e));

    private final String symbol;

    public static Optional<Operation> fromString(String symbol) {
        return Optional.ofNullable(stringToEnum.get(symbol));
    }

    @Override
    public String toString() {
        return symbol;
    }

    public abstract double apply(double x, double y);
}
